moving functions into _drainage, _imbibition, and _invasion files
